Re-estructured e2e tests for platform services
