+++ test/testdata/platform/services/dev/ephemeral/02-sonataflow_platform.yaml
