+++ test/testdata/platform/services/dev/ephemeral/kustomization.yaml
