+++ test/testdata/platform/services/dev/postgreSQL/02-sonataflow_platform.yaml
