+++ test/testdata/platform/services/prod/ephemeral/02-sonataflow_platform.yaml
