+++ test/testdata/platform/services/prod/ephemeral/kustomization.yaml
